feat: consolidates duplicative logic between update and updateByID operations
