+++ packages/payload/src/collections/operations/updateByID.ts
@@ -3,7 +3,6 @@ import type { DeepPartial } from 'ts-essentials'
 import httpStatus from 'http-status'
 
 import type { FindOneArgs } from '../../database/types.js'
-import type { Args } from '../../fields/hooks/beforeChange/index.js'
 import type { CollectionSlug } from '../../index.js'
 import type {
   PayloadRequest,
@@ -13,31 +12,20 @@ import type {
 } from '../../types/index.js'
 import type {
   Collection,
-  DataFromCollectionSlug,
   RequiredDataFromCollectionSlug,
   SelectFromCollectionSlug,
 } from '../config/types.js'
 
-import { ensureUsernameOrEmail } from '../../auth/ensureUsernameOrEmail.js'
 import executeAccess from '../../auth/executeAccess.js'
-import { generatePasswordSaltHash } from '../../auth/strategies/local/generatePasswordSaltHash.js'
 import { hasWhereAccessResult } from '../../auth/types.js'
 import { combineQueries } from '../../database/combineQueries.js'
 import { APIError, Forbidden, NotFound } from '../../errors/index.js'
-import { afterChange } from '../../fields/hooks/afterChange/index.js'
-import { afterRead } from '../../fields/hooks/afterRead/index.js'
-import { beforeChange } from '../../fields/hooks/beforeChange/index.js'
-import { beforeValidate } from '../../fields/hooks/beforeValidate/index.js'
-import { deleteAssociatedFiles } from '../../uploads/deleteAssociatedFiles.js'
 import { generateFileData } from '../../uploads/generateFileData.js'
-import { unlinkTempFiles } from '../../uploads/unlinkTempFiles.js'
-import { uploadFiles } from '../../uploads/uploadFiles.js'
-import { checkDocumentLockStatus } from '../../utilities/checkDocumentLockStatus.js'
 import { commitTransaction } from '../../utilities/commitTransaction.js'
 import { initTransaction } from '../../utilities/initTransaction.js'
 import { killTransaction } from '../../utilities/killTransaction.js'
 import { getLatestCollectionVersion } from '../../versions/getLatestCollectionVersion.js'
-import { saveVersion } from '../../versions/saveVersion.js'
+import { updateDocument } from './utilities/update.js'
 import { buildAfterOperation } from './utils.js'
 
 export type Arguments<TSlug extends CollectionSlug> = {
@@ -118,7 +106,7 @@ export const updateByIDOperation = async <
       throw new APIError('Missing ID of document to update.', httpStatus.BAD_REQUEST)
     }
 
-    let { data } = args
+    const { data } = args
     const password = data?.password
     const shouldSaveDraft = Boolean(draftArg && collectionConfig.versions.drafts)
     const shouldSavePassword = Boolean(password && collectionConfig.auth && !shouldSaveDraft)
@@ -159,43 +147,6 @@ export const updateByIDOperation = async <
     }
 
     // /////////////////////////////////////
-    // Handle potentially locked documents
-    // /////////////////////////////////////
-
-    await checkDocumentLockStatus({
-      id,
-      collectionSlug: collectionConfig.slug,
-      lockErrorMessage: `Document with ID ${id} is currently locked by another user and cannot be updated.`,
-      overrideLock,
-      req,
-    })
-
-    const originalDoc = await afterRead({
-      collection: collectionConfig,
-      context: req.context,
-      depth: 0,
-      doc: docWithLocales,
-      draft: draftArg,
-      fallbackLocale: null,
-      global: null,
-      locale,
-      overrideAccess: true,
-      req,
-      showHiddenFields: true,
-    })
-
-    if (args.collection.config.auth) {
-      ensureUsernameOrEmail<TSlug>({
-        authOptions: args.collection.config.auth,
-        collectionSlug: args.collection.config.slug,
-        data: args.data,
-        operation: 'update',
-        originalDoc,
-        req: args.req,
-      })
-    }
-
-    // /////////////////////////////////////
     // Generate data for all files and sizes
     // /////////////////////////////////////
 
@@ -209,265 +160,43 @@ export const updateByIDOperation = async <
       throwOnMissingFile: false,
     })
 
-    data = newFileData
+    // ///////////////////////////////////////////////
+    // Update document, runs all document level hooks
+    // ///////////////////////////////////////////////
 
-    // /////////////////////////////////////
-    // Delete any associated files
-    // /////////////////////////////////////
-
-    await deleteAssociatedFiles({
+    let result = await updateDocument<TSlug, TSelect>({
+      id,
+      accessResults,
+      autosave,
       collectionConfig,
       config,
-      doc: docWithLocales,
-      files: filesToUpload,
-      overrideDelete: false,
-      req,
-    })
-
-    // /////////////////////////////////////
-    // beforeValidate - Fields
-    // /////////////////////////////////////
-
-    data = await beforeValidate<DeepPartial<DataFromCollectionSlug<TSlug>>>({
-      id,
-      collection: collectionConfig,
-      context: req.context,
-      data,
-      doc: originalDoc,
-      global: null,
-      operation: 'update',
-      overrideAccess,
-      req,
-    })
-
-    // /////////////////////////////////////
-    // beforeValidate - Collection
-    // /////////////////////////////////////
-
-    await collectionConfig.hooks.beforeValidate.reduce(async (priorHook, hook) => {
-      await priorHook
-
-      data =
-        (await hook({
-          collection: collectionConfig,
-          context: req.context,
-          data,
-          operation: 'update',
-          originalDoc,
-          req,
-        })) || data
-    }, Promise.resolve())
-
-    // /////////////////////////////////////
-    // Write files to local storage
-    // /////////////////////////////////////
-
-    if (!collectionConfig.upload.disableLocalStorage) {
-      await uploadFiles(payload, filesToUpload, req)
-    }
-
-    // /////////////////////////////////////
-    // beforeChange - Collection
-    // /////////////////////////////////////
-
-    await collectionConfig.hooks.beforeChange.reduce(async (priorHook, hook) => {
-      await priorHook
-
-      data =
-        (await hook({
-          collection: collectionConfig,
-          context: req.context,
-          data,
-          operation: 'update',
-          originalDoc,
-          req,
-        })) || data
-    }, Promise.resolve())
-
-    // /////////////////////////////////////
-    // beforeChange - Fields
-    // /////////////////////////////////////
-
-    let publishedDocWithLocales = docWithLocales
-    let versionSnapshotResult
-
-    const beforeChangeArgs: Args<DataFromCollectionSlug<TSlug>> = {
-      id,
-      collection: collectionConfig,
-      context: req.context,
-      data: { ...data, id },
-      doc: originalDoc,
-      docWithLocales: undefined,
-      global: null,
-      operation: 'update',
-      req,
-      skipValidation:
-        shouldSaveDraft &&
-        collectionConfig.versions.drafts &&
-        !collectionConfig.versions.drafts.validate &&
-        data._status !== 'published',
-    }
-
-    if (publishSpecificLocale) {
-      versionSnapshotResult = await beforeChange({
-        ...beforeChangeArgs,
-        docWithLocales,
-      })
-
-      const lastPublished = await getLatestCollectionVersion({
-        id,
-        config: collectionConfig,
-        payload,
-        published: true,
-        query: findOneArgs,
-        req,
-      })
-
-      publishedDocWithLocales = lastPublished ? lastPublished : {}
-    }
-
-    let result = await beforeChange({
-      ...beforeChangeArgs,
-      docWithLocales: publishedDocWithLocales,
-    })
-
-    // /////////////////////////////////////
-    // Handle potential password update
-    // /////////////////////////////////////
-
-    const dataToUpdate: Record<string, unknown> = { ...result }
-
-    if (shouldSavePassword && typeof password === 'string') {
-      const { hash, salt } = await generatePasswordSaltHash({
-        collection: collectionConfig,
-        password,
-        req,
-      })
-      dataToUpdate.salt = salt
-      dataToUpdate.hash = hash
-      delete dataToUpdate.password
-      delete data.password
-    }
-
-    // /////////////////////////////////////
-    // Update
-    // /////////////////////////////////////
-
-    if (!shouldSaveDraft || data._status === 'published') {
-      result = await req.payload.db.updateOne({
-        id,
-        collection: collectionConfig.slug,
-        data: dataToUpdate,
-        locale,
-        req,
-        select,
-      })
-    }
-
-    // /////////////////////////////////////
-    // Create version
-    // /////////////////////////////////////
-
-    if (collectionConfig.versions) {
-      result = await saveVersion({
-        id,
-        autosave,
-        collection: collectionConfig,
-        docWithLocales: result,
-        draft: shouldSaveDraft,
-        payload,
-        publishSpecificLocale,
-        req,
-        select,
-        snapshot: versionSnapshotResult,
-      })
-    }
-
-    // /////////////////////////////////////
-    // afterRead - Fields
-    // /////////////////////////////////////
-
-    result = await afterRead({
-      collection: collectionConfig,
-      context: req.context,
+      data: newFileData,
       depth,
-      doc: result,
-      draft: draftArg,
+      docWithLocales,
+      draftArg,
       fallbackLocale,
-      global: null,
+      filesToUpload,
       locale,
       overrideAccess,
+      overrideLock,
+      payload,
       populate,
+      publishSpecificLocale,
       req,
       select,
       showHiddenFields,
     })
 
     // /////////////////////////////////////
-    // afterRead - Collection
-    // /////////////////////////////////////
-
-    await collectionConfig.hooks.afterRead.reduce(async (priorHook, hook) => {
-      await priorHook
-
-      result =
-        (await hook({
-          collection: collectionConfig,
-          context: req.context,
-          doc: result,
-          req,
-        })) || result
-    }, Promise.resolve())
-
-    // /////////////////////////////////////
-    // afterChange - Fields
-    // /////////////////////////////////////
-
-    result = await afterChange({
-      collection: collectionConfig,
-      context: req.context,
-      data,
-      doc: result,
-      global: null,
-      operation: 'update',
-      previousDoc: originalDoc,
-      req,
-    })
-
-    // /////////////////////////////////////
-    // afterChange - Collection
-    // /////////////////////////////////////
-
-    await collectionConfig.hooks.afterChange.reduce(async (priorHook, hook) => {
-      await priorHook
-
-      result =
-        (await hook({
-          collection: collectionConfig,
-          context: req.context,
-          doc: result,
-          operation: 'update',
-          previousDoc: originalDoc,
-          req,
-        })) || result
-    }, Promise.resolve())
-
-    // /////////////////////////////////////
     // afterOperation - Collection
     // /////////////////////////////////////
 
-    result = await buildAfterOperation({
+    result = (await buildAfterOperation({
       args,
       collection: collectionConfig,
       operation: 'updateByID',
       result,
-    })
-
-    await unlinkTempFiles({
-      collectionConfig,
-      config,
-      req,
-    })
+    })) as TransformCollectionWithSelect<TSlug, TSelect>
 
     // /////////////////////////////////////
     // Return results
@@ -477,7 +206,7 @@ export const updateByIDOperation = async <
       await commitTransaction(req)
     }
 
-    return result as TransformCollectionWithSelect<TSlug, TSelect>
+    return result
   } catch (error: unknown) {
     await killTransaction(args.req)
     throw error
